1.将原先的url数组替换成api接口 2.将项目解耦成MPV模式 3.采用瀑布流+卡片控件的布局方式，在图片下方显示图片信息，并支持卡片拖拽动作 4.原本图片压缩成正方形尺寸，现改为按比例缩放 5.将原来的HttpUrlConnection替换成OkHttpConnection，并增加对连接超时的处理 6.更改线程池配置，为了适应IO密集型工作，使更多线程能响应IO请求，将核心线程数增加到2*CPU，总线程增加到Integer.MAX_VALUE，阻塞队列长度限制为10，线程闲置时间减小为5

+++ app/src/main/java/com/example/zhaolexi/imageloader/utils/MyUtils.java
@@ -1,18 +1,17 @@
 package com.example.zhaolexi.imageloader.utils;
 
-import java.io.Closeable;
-import java.io.IOException;
-import java.util.List;
-
 import android.app.ActivityManager;
 import android.app.ActivityManager.RunningAppProcessInfo;
 import android.content.Context;
-import android.content.res.TypedArray;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.util.DisplayMetrics;
 import android.util.TypedValue;
 import android.view.WindowManager;
+
+import java.io.Closeable;
+import java.io.IOException;
+import java.util.List;
 
 public class MyUtils {
 
@@ -48,8 +47,8 @@ public class MyUtils {
         return dm;
     }
 
-    public static float dp2px(Context context, float dp) {
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp,
+    public static int dp2px(Context context, int dp) {
+        return (int)TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp,
                 context.getResources().getDisplayMetrics());
     }
